function to compose filepath for newfile

+++ data_database.py
@@ -65,7 +65,7 @@ class Database(object):
     def get_watched_folders(self):
         c = self.get_cursor()
         watched_folders = [s for s in c.execute('select path, experiment from watched_folders')]
-        interval = int(self.get_setting('watch_folder_interval'))
+        interval = float(self.get_setting('watch_folder_interval'))
         watched_folders = map(lambda a: (a[0], interval, a[1]), watched_folders)
         if type(watched_folders) is list:
             return watched_folders
